perf(ui): memoize markdown parsing and per-render detail work
Markdown re-parsed its text on every parent render in the 30 files that mount it,
including every transcript row: 21 parses for 21 renders of the same text, now 1.
The parser moves to a sibling module so the cost is measurable, and the component
is memo-wrapped on three primitive props. resolveDetailFields is hoisted out of
JSX in the issue, PR pane and PR studio panels, resolverLaneEntries is memoized in
LensColumn, and the transcripts, agentRunHistory and agentTurnState whole-slice
subscriptions become per-agent selections.
